[sdk] Add web preview support (#17)
This PR adds auto-configuration and a transport for using the Snack web-player (v2). This code was previously implemented in snack-web and also in the example app. This consolidates the code in the snack-sdk and make it easy for anyone to use the web-player. This also provides a better upgrade path when we want to provide access to partners.