[sdk] Fix dependencies object modification & scoped dependency names (#37)
Fixes https://forums.expo.io/t/cannot-open-old-snack/47209/3, which was caused by the snack-sdk modifying the initial dependencies object and setting an error status on the faulty dependency. This in turn caused the __INITIAL_DATA obect to be modified and serialised to fail because the Error object could not be serialised.

The faulty dependency was indeed not really invalid. It was a scoped dependency name with a subpath, but that particular case was not correctly supported.