[website] Add website app (#42)
Moves the Snack web-app into the expo/snack monorepo.

After this PR, the website folder contains the Snack web-app. This also adds Github Actions which build the Docker image for staging and production upon creating a Pull Request or pushing to main. When pushed to main, the staging github action will build and deploy the docker image to staging.snack.expo.io. The production docker-image is also build upon pushing to main but requires a manual "deploy" confirmation to trigger deploy to snack.expo.io.