[tools] Replace PM2 with terminal app (#100)
PM2 has proved to be too "heavy" and doesn't work on Windows currently. It also required duplication of env configuration which make more complex to maintain. Additionally, PM2 logging was flaky and PM2 very inconveniently cleared the logs when switching to another terminal.

This PR removes PM2 and replaces it with a basic terminal app based on React Ink. It can be easily extended to support more services like Snackager and the Web-player runtime.