[website] Fix footer elements appearance while in error mode (#99)
This PR adds few styles overwrites for states like hover and active to retain readability even in error mode. This probably can be improved in future utilizing the hover color with an alpha value (transparency).

Additionally in this PR I have fixed the select arrow character position and I have updated the select background so it's consistent with latest tweak for the Togglers.

Last small change fixes the Toast button colors in Dark Mode.