[snackager] Update test snapshots (#167)
Refreshes the `.lock` files which causes the latest @babel/runtime to be used. This causes the tests to closer represent the current reality when bundling. This PR also sorts the reported externals to prevent future test-cases from failing when the generated bundle is slightly different (causing a different import ordering).