[runtime] Replace non-deterministic yarn behavior with two same packages (#273)
Instead of letting yarn install two separate instances of a package, we copy the plugin before applying a patch to it. That way, we can use a version in the snack runtime, and one in the snack code itself.

We might want to look for alternatives to this issue though.