Fix initializing embedded snacks using `data-snack-files`
`data-snack-files` allows specifying specific filenames for code, instead of `data-snack-code` which is placed as `App.js`, or `data-snack-id` which points to a saved Snack. `data-snack-files` replaces `data-snack-code` if present, but an element with only the former will not be initialized.

This adds `data-snack-files` to the list of searched for elements, to catch embedded snacks not using `data-snack-code` (e.g. to embed TypeScript instead of JavaScript).