Fix syntax highlighting for embedded tsx files
Syntax highlighting for embedded snacks using tsx is poor in comparison to jsx.

I haven't tried to test this, but looking through to find where syntax highlighting was happening, I noticed that `SimpleEditor` was setting a language of 'typescript' for tsx files, instead of 'tsx' https://github.com/PrismJS/prism/blob/master/components/prism-tsx.js This makes it so that tsx files in SimpleEditor set the correct language.

I am not familiar with this code, so please let me know if this isn't the right place for embedded snacks, or if there is a way to easily test.